More UI changes, Alerts now its own component
