+++ Website/linkbus-website/src/components/Option.vue
@@ -1,0 +1,13 @@
+<template>
+    
+</template>
+
+<script>
+    export default {
+        name: "Option"
+    }
+</script>
+
+<style scoped>
+
+</style>
